+ Added SCharacterInventory struct. + Item descriptions are not stored in SInventorySlot anymore and now we just use the description index for accessing the item descriptions. + SCharacterScore members are now uint64_t instead of int * Fixed text reporting the combat points at the beginning of the player's turn. * Fixed some missing casts for 32 bit builds.

+++ KitsuRPG/Menu.h
@@ -22,7 +22,7 @@ int displayMenu(size_t optionCount, const std::string& title, const SMenuItem (&
 		printf(	"Make your selection:\n" );
 		
 		// Print menu options
-		for(int i=0; i<optionCount; i++)
+		for(size_t i=0; i<optionCount; i++)
 			printf("%u: %s.\n", i+1, menuItems[i].Text.c_str());	
 		
 		// Get user input
